+++ apps/nuxt/server/types/auth.d.ts
@@ -3,6 +3,7 @@ declare module 'nuxt-auth-utils' {
     id: string
     email: string
     name: string | null
+    providers?: string[]
   }
   interface UserSession {
     user: User